+++ packages/core/src/components/layout/spacer.ts
@@ -1,23 +1,23 @@
 // Internal Imports
-import { Component } from "../component";
-import { DOMUtils } from "../../services";
-import * as Configuration from "../../configuration";
+import { Component } from '../component';
+import { DOMUtils } from '../../services';
+import * as Configuration from '../../configuration';
 
 export class Spacer extends Component {
-	type = "spacer";
+	type = 'spacer';
 
 	render() {
-		DOMUtils.appendOrSelect(this.getContainerSVG(), "rect")
-			.attr("x", 0)
-			.attr("y", 0)
+		DOMUtils.appendOrSelect(this.getContainerSVG(), 'rect')
+			.attr('x', 0)
+			.attr('y', 0)
 			.attr(
-				"width",
+				'width',
 				this.configs.size || Configuration.spacers.default.size
 			)
 			.attr(
-				"height",
+				'height',
 				this.configs.size || Configuration.spacers.default.size
 			)
-			.attr("opacity", 0);
+			.attr('opacity', 0);
 	}
 }