+++ packages/core/src/services/curves.ts
@@ -21,14 +21,14 @@ import {
 	CurveBundleFactory,
 	CurveCardinalFactory,
 	CurveCatmullRomFactory,
-	CurveGenerator
-} from "d3-shape";
+	CurveGenerator,
+} from 'd3-shape';
 
 // this import is needed because of typescript bug (error TS4029)
-import { Path } from "d3-path";
+import { Path } from 'd3-path';
 
 // Internal Imports
-import { Service } from "./service";
+import { Service } from './service';
 
 export class Curves extends Service {
 	curveTypes = {
@@ -49,17 +49,17 @@ export class Curves extends Service {
 		curveNatural,
 		curveStep,
 		curveStepAfter,
-		curveStepBefore
+		curveStepBefore,
 	};
 
 	getD3Curve() {
-		let curveName = "curveLinear";
+		let curveName = 'curveLinear';
 		const curveOptions = this.model.getOptions().curve;
 
 		// Parse curve type whether the user provided a string
 		// Or an object with more options
 		if (curveOptions) {
-			if (typeof curveOptions === "string") {
+			if (typeof curveOptions === 'string') {
 				// curve: 'string'
 				curveName = curveOptions;
 			} else {
@@ -87,6 +87,6 @@ export class Curves extends Service {
 		console.warn(
 			`The curve type '${curveName}' is invalid, using 'curveLinear' instead`
 		);
-		return this.curveTypes["curveLinear"];
+		return this.curveTypes['curveLinear'];
 	}
 }