+++ CHANGELOG.md
@@ -3,6 +3,65 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.41.22](https://github.com/carbon-design-system/carbon-charts/compare/v0.41.21...v0.41.22) (2021-01-15)
+
+**Note:** Version bump only for package @carbon/charts-monorepo
+
+
+
+
+
+## [0.41.21](https://github.com/carbon-design-system/carbon-charts/compare/v0.41.20...v0.41.21) (2021-01-13)
+
+**Note:** Version bump only for package @carbon/charts-monorepo
+
+
+
+
+
+## [0.41.20](https://github.com/carbon-design-system/carbon-charts/compare/v0.41.19...v0.41.20) (2021-01-13)
+
+**Note:** Version bump only for package @carbon/charts-monorepo
+
+
+
+
+
+## [0.41.19](https://github.com/carbon-design-system/carbon-charts/compare/v0.41.18...v0.41.19) (2021-01-13)
+
+**Note:** Version bump only for package @carbon/charts-monorepo
+
+
+
+
+
+## [0.41.18](https://github.com/carbon-design-system/carbon-charts/compare/v0.41.17...v0.41.18) (2021-01-13)
+
+**Note:** Version bump only for package @carbon/charts-monorepo
+
+
+
+
+
+## [0.41.17](https://github.com/carbon-design-system/carbon-charts/compare/v0.41.16...v0.41.17) (2021-01-13)
+
+**Note:** Version bump only for package @carbon/charts-monorepo
+
+
+
+
+
+## [0.41.16](https://github.com/carbon-design-system/carbon-charts/compare/v0.41.15...v0.41.16) (2021-01-13)
+
+
+### Bug Fixes
+
+* **core:** add number to the donut center interface ([#929](https://github.com/carbon-design-system/carbon-charts/issues/929)) ([2f012a7](https://github.com/carbon-design-system/carbon-charts/commit/2f012a7915c3250b1044016330b91bc4e76206d8))
+
+
+
+
+
 ## [0.41.15](https://github.com/carbon-design-system/carbon-charts/compare/v0.41.14...v0.41.15) (2021-01-07)